pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ladybug-viz"
version = "0.1.0"
description = "A reusable Django app for visualizing LadybugDB graph databases"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "django-ninja>=1.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "real-ladybug>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
]

[tool.hatch.build.targets.wheel]
packages = ["ladybug_viz"]
